Manifest is now divided into separate files for representing Cargo.toml and package.json files, for Cargo and npm respectively.
