+++ src/manifest/mod.rs
@@ -8,120 +8,11 @@ use serde_json;
 use toml;
 use PBAR;
 
-#[derive(Deserialize)]
-pub struct CargoManifest {
-    package: CargoPackage,
-}
+mod cargo;
+mod npm;
 
-impl CargoManifest {
-    pub fn get_crate_name(&self) -> String {
-        self.package.name.clone()
-    }
-}
-
-#[derive(Deserialize)]
-pub struct CargoPackage {
-    name: String,
-    authors: Vec<String>,
-    description: Option<String>,
-    version: String,
-    license: Option<String>,
-    repository: Option<String>,
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct NpmPackage {
-    name: String,
-    collaborators: Vec<String>,
-    description: Option<String>,
-    version: String,
-    license: Option<String>,
-    repository: Option<Repository>,
-    files: Vec<String>,
-    main: String,
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct Repository {
-    #[serde(rename = "type")]
-    ty: String,
-    url: String,
-}
-
-#[derive(Serialize)]
-struct PackageFiles {
-    main: String,
-    files: Vec<String>,
-}
-
-impl NpmPackage {
-    /// Create an NpmPackage from a borrowed `CargoManifest` object.
-    pub fn new(cargo: &CargoManifest, scope: Option<String>) -> NpmPackage {
-        let name = NpmPackage::scoped_name(&cargo.package.name, scope);
-        let repository = NpmPackage::get_repo(cargo);
-        let PackageFiles { files, main } = NpmPackage::get_filenames(cargo);
-        NpmPackage {
-            name,
-            collaborators: cargo.package.authors.clone(),
-            description: cargo.package.description.clone(),
-            version: cargo.package.version.clone(),
-            license: cargo.package.license.clone(),
-            repository,
-            files,
-            main,
-        }
-    }
-
-    fn scoped_name(name: &str, scope: Option<String>) -> String {
-        match scope {
-            Some(s) => format!("@{}/{}", s, name),
-            None => name.to_string(),
-        }
-    }
-
-    pub fn name(&self) -> String {
-        self.name.clone()
-    }
-
-    pub fn repository_url(&self) -> Option<String> {
-        match &self.repository {
-            Some(repo) => Some(repo.url.clone()),
-            None => None,
-        }
-    }
-
-    pub fn repository_type(&self) -> Option<String> {
-        match &self.repository {
-            Some(repo) => Some(repo.ty.clone()),
-            None => None,
-        }
-    }
-
-    pub fn main(&self) -> String {
-        self.main.clone()
-    }
-
-    pub fn files(&self) -> Vec<String> {
-        self.files.clone()
-    }
-
-    fn get_repo(cargo: &CargoManifest) -> Option<Repository> {
-        cargo.package.repository.clone().map(|repo_url| Repository {
-            ty: "git".to_string(),
-            url: repo_url,
-        })
-    }
-
-    fn get_filenames(cargo: &CargoManifest) -> PackageFiles {
-        let filename = cargo.package.name.clone().replace("-", "_");
-        let js_file = format!("{}.js", filename);
-        let wasm_file = format!("{}_bg.wasm", filename);
-        PackageFiles {
-            main: js_file,
-            files: vec![wasm_file],
-        }
-    }
-}
+pub use self::cargo::{CargoManifest, CargoPackage};
+pub use self::npm::{NpmPackage, Repository};
 
 /// Try to read the `Cargo.toml` file in the given directory.
 pub fn read_cargo_toml(path: &str) -> Result<CargoManifest, Error> {
